drivers/intel/fsp2_0: Adjust framebuffer_bar type to uint64_t
With RESOURCE_ALLOCATION_TOP_DOWN enabled, the framebuffer BAR may be
allocated above 4G. If coreboot is built for IA32, this means that the
BAR would be truncated to 32 bits when writing framebuffer info to
CBMEM, and graphics would not work in the payload.

Set the BAR size to 64 bits to avoid this condition.

+++ src/drivers/intel/fsp2_0/include/fsp/graphics.h
@@ -11,7 +11,7 @@
  * Must be called after PCI enumeration to make sure that the BAR
  * doesn't change any more.
  */
-void fsp_report_framebuffer_info(const uintptr_t framebuffer_bar,
+void fsp_report_framebuffer_info(const uint64_t framebuffer_bar,
 				 enum lb_fb_orientation orientation);
 
 /* SoC Overrides */
